Refactor balance fetching
