+++ app/src/marketplace/containers/GlobalInfoWatcher/index.jsx
@@ -18,6 +18,7 @@ import { useBalances } from '$shared/hooks/useBalances'
 import { selectUserData } from '$shared/modules/user/selectors'
 import type { NumberString } from '$shared/flowtype/common-types'
 import useEthereumIdentities from '$shared/modules/integrationKey/hooks/useEthereumIdentities'
+import { isEthereumAddress } from '$mp/utils/validate'
 
 type Props = {
     children?: Node,
@@ -112,13 +113,10 @@ export const GlobalInfoWatcher = ({ children }: Props) => {
     const { username } = user || {}
 
     useEffect(() => {
-        if (!username) { return () => {} }
+        if (!username || !isEthereumAddress(username)) { return () => {} }
 
         // fetch the integration keys first and then start polling for the balance
-        loadIntegrationKeys()
-            .then(() => {
-                balancePoll()
-            })
+        balancePoll()
 
         return () => {
             clearTimeout(balanceTimeout.current)
