+++ app/src/shared/flowtype/integration-key-types.js
@@ -1,6 +1,5 @@
 // @flow
 
-import type { NumberString } from '$shared/flowtype/common-types'
 import type { Address } from './web3-types'
 
 export type IntegrationKeyId = string
@@ -26,19 +25,6 @@ export type Challenge = {
     challenge: string,
     expires: Date,
     id: string,
-}
-
-export const BalanceType = {
-    ETH: 'ETH',
-    DATA: 'DATA',
-}
-
-export type Balances = {
-    [$Values<typeof BalanceType>]: NumberString,
-}
-
-export type BalanceList = {
-    [Address]: Balances,
 }
 
 export type Account = {
