+++ app/src/shared/flowtype/user-types.js
@@ -1,5 +1,7 @@
 // @flow
 
+import type { NumberString } from '$shared/flowtype/common-types'
+
 export type UserId = string
 
 export type User = {
@@ -9,4 +11,13 @@ export type User = {
     username: string,
     imageUrlSmall: string,
     imageUrlLarge: string,
+}
+
+export const BalanceType = {
+    ETH: 'ETH',
+    DATA: 'DATA',
+}
+
+export type Balances = {
+    [$Values<typeof BalanceType>]: NumberString,
 }
