+++ app/src/shared/hooks/useBalances.js
@@ -3,10 +3,8 @@
 import { useCallback, useMemo } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 
-import type { Address } from '$shared/flowtype/web3-types'
-import { BalanceType } from '$shared/flowtype/integration-key-types'
-import { selectBalances } from '$shared/modules/integrationKey/selectors'
-import { updateBalances } from '$shared/modules/integrationKey/actions'
+import { selectBalances } from '$shared/modules/user/selectors'
+import { updateBalances } from '$shared/modules/user/actions'
 
 export function useBalances() {
     const dispatch = useDispatch()
@@ -26,16 +24,4 @@ export function useBalances() {
         update,
         balances,
     ])
-}
-
-export function useAccountBalance(account: Address, balanceType: $Values<typeof BalanceType>) {
-    const { balances } = useBalances()
-
-    const balance = useMemo(() => {
-        const accountBalance = balances[account] || {}
-
-        return accountBalance[balanceType] ? accountBalance[balanceType].decimalPlaces(4).toString() : '-'
-    }, [balances, account, balanceType])
-
-    return balance
 }
