+++ app/src/shared/modules/integrationKey/reducer.js
@@ -8,7 +8,6 @@ import type {
     IntegrationKeysAction,
     IntegrationKeyIdAction,
     IntegrationKeysErrorAction,
-    SetBalanceAction,
 } from './types'
 import {
     INTEGRATION_KEYS_REQUEST,
@@ -20,7 +19,6 @@ import {
     DELETE_INTEGRATION_KEY_REQUEST,
     DELETE_INTEGRATION_KEY_SUCCESS,
     DELETE_INTEGRATION_KEY_FAILURE,
-    SET_BALANCE,
 } from './constants'
 
 export const initialState: IntegrationKeyState = {
@@ -31,7 +29,6 @@ export const initialState: IntegrationKeyState = {
     creatingIdentityError: null,
     removingIntegrationKey: false,
     removingIntegrationError: null,
-    balances: {},
 }
 
 const reducer: (IntegrationKeyState) => IntegrationKeyState = handleActions({
@@ -87,17 +84,6 @@ const reducer: (IntegrationKeyState) => IntegrationKeyState = handleActions({
         removingIntegrationKey: false,
         removingIntegrationError: action.payload.error,
     }),
-
-    [SET_BALANCE]: (state: IntegrationKeyState, action: SetBalanceAction) => ({
-        ...state,
-        balances: {
-            ...state.balances,
-            [action.payload.account]: {
-                ...action.payload.balances,
-            },
-        },
-    }),
-
 }, initialState)
 
 export default reducer
