+++ app/src/shared/modules/integrationKey/selectors.js
@@ -4,7 +4,7 @@ import { createSelector } from 'reselect'
 import { denormalize } from 'normalizr'
 
 import type { IntegrationKeyState, StoreState, EntitiesState } from '$shared/flowtype/store-state'
-import type { IntegrationKeyIdList, IntegrationKeyList, BalanceList } from '$shared/flowtype/integration-key-types'
+import type { IntegrationKeyIdList, IntegrationKeyList } from '$shared/flowtype/integration-key-types'
 import type { ErrorInUi } from '$shared/flowtype/common-types'
 import { selectEntities } from '$shared/modules/entities/selectors'
 import { integrationKeysSchema } from '$shared/modules/entities/schema'
@@ -42,23 +42,3 @@ export const selectCreatingIdentityError: (StoreState) => ?ErrorInUi = createSel
     (subState: IntegrationKeyState): ?ErrorInUi => subState.creatingIdentityError,
 )
 
-export const selectBalances: (StoreState) => BalanceList = createSelector(
-    selectIntegrationKeyState,
-    selectEthereumIdentities,
-    (subState: IntegrationKeyState, ethIdentities: IntegrationKeyList): BalanceList => {
-        const uniqueAccounts = [...(new Set([
-            ...(ethIdentities || []).map(({ json }) => json.address).filter(Boolean),
-        ]))]
-
-        return uniqueAccounts.reduce((result, address) => {
-            if (!subState.balances[address]) {
-                return result
-            }
-
-            return ({
-                ...result,
-                [address]: subState.balances[address],
-            })
-        }, {})
-    },
-)
