+++ app/src/shared/modules/integrationKey/services.js
@@ -3,7 +3,6 @@
 import { get, post, del, put } from '$shared/utils/api'
 import type { ApiResult } from '$shared/flowtype/common-types'
 import {
-    BalanceType,
     type IntegrationKeyId,
     type IntegrationKey,
     type Challenge,
@@ -11,7 +10,6 @@ import {
 } from '$shared/flowtype/integration-key-types'
 import type { Address, Hash } from '$shared/flowtype/web3-types'
 import { integrationKeyServices } from '$shared/utils/constants'
-import { getDataTokenBalance, getEthBalance } from '$mp/utils/web3'
 
 import {
     ChallengeFailedError,
@@ -90,26 +88,6 @@ export const deleteIntegrationKey = (id: IntegrationKeyId): ApiResult<null> => d
         id,
     }),
 })
-
-type GetBalance = {
-    address: Address,
-    type: $Values<typeof BalanceType>,
-    usePublicNode?: boolean,
-}
-
-export async function getBalance({ address, type, usePublicNode = false }: GetBalance) {
-    let balance
-
-    if (type === BalanceType.ETH) {
-        balance = await getEthBalance(address, usePublicNode)
-    } else if (type === BalanceType.DATA) {
-        balance = await getDataTokenBalance(address, usePublicNode)
-    } else {
-        throw new Error('Unknown balance type!')
-    }
-
-    return balance
-}
 
 export const getEnsDomains = ({ addresses }: {
     addresses: Array<Address>,
