+++ app/src/shared/modules/integrationKey/types.js
@@ -1,8 +1,7 @@
 // @flow
 
 import type { PayloadAction, ErrorInUi } from '$shared/flowtype/common-types'
-import type { IntegrationKeyId, IntegrationKeyIdList, Balances } from '$shared/flowtype/integration-key-types'
-import type { Address } from '$shared/flowtype/web3-types'
+import type { IntegrationKeyId, IntegrationKeyIdList } from '$shared/flowtype/integration-key-types'
 
 export type IntegrationKeysAction = PayloadAction<{
     ethereumIdentities: IntegrationKeyIdList,
@@ -19,10 +18,3 @@ export type IntegrationKeyIdAction = PayloadAction<{
 }>
 
 export type IntegrationKeyIdActionCreator = (IntegrationKeyId) => IntegrationKeyIdAction
-
-export type SetBalanceAction = PayloadAction<{
-    account: Address,
-    balances: Balances,
-}>
-
-export type SetBalanceActionCreator = (Address, Balances) => SetBalanceAction
