+++ app/src/shared/modules/user/constants.js
@@ -19,3 +19,5 @@ export const UPDATE_AVATAR_FAILURE = 'shared/user/UPDATE_AVATAR_FAILURE'
 export const DELETE_USER_ACCOUNT_REQUEST = 'shared/user/DELETE_USER_ACCOUNT_REQUEST'
 export const DELETE_USER_ACCOUNT_SUCCESS = 'shared/user/DELETE_USER_ACCOUNT_SUCCESS'
 export const DELETE_USER_ACCOUNT_FAILURE = 'shared/user/DELETE_USER_ACCOUNT_FAILURE'
+
+export const SET_BALANCE = 'shared/user/SET_BALANCE'
