+++ app/src/shared/modules/user/types.js
@@ -1,7 +1,7 @@
 // @flow
 
 import type { PayloadAction, ErrorInUi } from '$shared/flowtype/common-types'
-import type { User } from '$shared/flowtype/user-types'
+import type { User, Balances } from '$shared/flowtype/user-types'
 
 export type UserDataAction = PayloadAction<{
     user: User
@@ -12,3 +12,9 @@ export type UserErrorAction = PayloadAction<{
     error: ErrorInUi,
 }>
 export type UserErrorActionCreator = (error: ErrorInUi) => UserErrorAction
+
+export type SetBalanceAction = PayloadAction<{
+    balances: Balances,
+}>
+
+export type SetBalanceActionCreator = (Balances) => SetBalanceAction
