+++ app/src/shared/test/modules/user/actions.test.js
@@ -10,6 +10,7 @@ describe('user - actions', () => {
     beforeEach(() => {
         oldStreamrApiUrl = process.env.STREAMR_API_URL
         process.env.STREAMR_API_URL = ''
+        jest.spyOn(services, 'getBalance').mockImplementation(jest.fn())
     })
 
     afterEach(() => {
