+++ app/src/shared/test/modules/user/services.test.js
@@ -1,6 +1,8 @@
 import moxios from 'moxios'
 
 import * as services from '$shared/modules/user/services'
+import * as utils from '$mp/utils/web3'
+import { BalanceType } from '$shared/flowtype/user-types'
 
 describe('user - services', () => {
     let dateNowSpy
@@ -125,6 +127,45 @@ describe('user - services', () => {
 
             const result = await services.uploadProfileAvatar(imageToUpload)
             expect(result).toStrictEqual(data)
+        })
+    })
+
+    describe('getBalance', () => {
+        it('gets ETH balance', async () => {
+            jest.spyOn(utils, 'getEthBalance').mockImplementation(jest.fn(() => '123'))
+
+            const balance = await services.getBalance({
+                address: 'testAccount',
+                type: BalanceType.ETH,
+            })
+
+            expect(balance).toBe('123')
+        })
+
+        it('gets token balance', async () => {
+            jest.spyOn(utils, 'getDataTokenBalance').mockImplementation(jest.fn(() => '123'))
+
+            const balance = await services.getBalance({
+                address: 'testAccount',
+                type: BalanceType.DATA,
+            })
+            expect(balance).toBe('123')
+        })
+
+        it('throws an error if type is unknown', async () => {
+            let balance
+            let error
+            try {
+                balance = await services.getBalance({
+                    adress: 'testAccount',
+                    type: 'someToken',
+                })
+            } catch (e) {
+                error = e
+            }
+
+            expect(error).toBeDefined()
+            expect(balance).not.toBeDefined()
         })
     })
 })
